refactor topic list component
