+++ src/components/forums/forum/topic-list/index.ts
@@ -1,0 +1,1 @@
+export { TopicList } from './TopicList';
