+++ src/components/forums/index.tsx
@@ -8,8 +8,7 @@ export * from './forum/EditForum';
 export * from './forum/ForumContent';
 export * from './forum/ManageModerators';
 export * from './forum/ManageOwners';
-export * from './forum/TopicList';
-export * from './forum/TopicListRow';
+export * from './forum/topic-list';
 export * from './forum/UploadForumBanner';
 
 export * from './topic/CreatePost';
